Mandate use of the signature hash algorithm
As discussed on list.

I also took the liberty to explain why this is useful, and to change the last sentence.  The original statement (operating in an environment that requires stronger hashes, means that you know that other endpoints will support the stronger hash; however, you might have local policy that requires the stronger hash, meaning that you accept the possibility that the connection will fail if the peer doesn't support the stronger hash).